+++ src/state_idx.h
@@ -6,9 +6,9 @@
 struct imm_state;
 struct state_idx;
 
-void state_idx_add(struct state_idx* state_idx, struct imm_state const* state, uint32_t idx);
+void state_idx_add(struct state_idx* state_idx, struct imm_state const* state, uint16_t idx);
 struct state_idx* state_idx_create(void);
 void              state_idx_destroy(struct state_idx* state_idx);
-uint32_t          state_idx_find(struct state_idx const* state_idx, struct imm_state const* state);
+uint16_t          state_idx_find(struct state_idx const* state_idx, struct imm_state const* state);
 
 #endif
